Verbesserung in Create a Simple Rail Network Simulation
